+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.2.0
+- Changed: ModelConvention marked as Flags
+- Changed: For ModelAttribute added targets Enum and Interface
+
 # 0.1.0
 - Added ModelAttribute and ModelConvention